Use double more

+++ src/support/Location.hpp
@@ -6,15 +6,15 @@ namespace Lattice {
 
 class Location final {
 public:
-    glm::vec3 scale {1};
-    glm::vec3 rotation {0};
-    glm::vec3 position {0};
+    glm::dvec3 scale {1};
+    glm::dvec3 rotation {0};
+    glm::dvec3 position {0};
 
-    glm::mat4 matrix() const;
+    glm::dmat4 matrix() const;
 
-    glm::vec3 right() const;
-    glm::vec3 up() const;
-    glm::vec3 backward() const;
+    glm::dvec3 right() const;
+    glm::dvec3 up() const;
+    glm::dvec3 backward() const;
 
     bool operator==(const Location& other) const;
     bool operator!=(const Location& other) const;
